feat: Add NTP time synchronization state
Implements a new state to synchronize the system time with an NTP server after establishing a Wi-Fi connection. The current time is then printed to the console every second in the idle state.

This provides the necessary foundation for calendar event processing.

TODO: The current implementation does not set a timezone, resulting in the time being displayed in UTC. A future commit should add a state to automatically determine the user's timezone from their IP address after connecting to Wi-Fi.

verified: Bo Lin <k2345777@gmail.com>

Signed-off-by: GEMINI <gemini@google.com>